+++ app/client/src/pages/AppIDE/layouts/routers/QueryEditor/tests/ApiRoutes.test.tsx
@@ -1,0 +1,220 @@
+import React from "react";
+import { Route } from "react-router-dom";
+import { render } from "test/testUtils";
+import IDE from "../../../index";
+import { createMessage, EDITOR_PANE_TEXTS } from "ee/constants/messages";
+import { BUILDER_PATH } from "ee/constants/routes/appRoutes";
+import { EditorEntityTab, EditorViewMode } from "IDE/Interfaces/EditorTypes";
+import { APIFactory } from "test/factories/Actions/API";
+import { getIDETestState } from "test/factories/AppIDEFactoryUtils";
+import { PageFactory } from "test/factories/PageFactory";
+import { waitFor } from "@testing-library/react";
+
+describe("IDE URL rendering of Queries", () => {
+  describe("API Routes", () => {
+    it("Renders Api routes in Full screen", async () => {
+      const page = PageFactory.build();
+      const anApi = APIFactory.build({ pageId: page.pageId });
+      const state = getIDETestState({
+        pages: [page],
+        actions: [anApi],
+        tabs: {
+          [EditorEntityTab.QUERIES]: [anApi.baseId],
+          [EditorEntityTab.JS]: [],
+        },
+      });
+
+      const { getAllByRole, getAllByText, getByRole, getByTestId } = render(
+        <Route path={BUILDER_PATH}>
+          <IDE />
+        </Route>,
+        {
+          url: `/app/applicationSlug/pageSlug-${page.basePageId}/edit/api/${anApi.baseId}`,
+          initialState: state,
+        },
+      );
+
+      await waitFor(
+        async () => {
+          const elements = getAllByText("Api1"); // Use the common test ID or selector
+
+          expect(elements).toHaveLength(2); // Wait until there are exactly 3 elements
+        },
+        { timeout: 3000, interval: 500 },
+      );
+
+      // There will be 2 Api1 text (Left pane list, editor tab)
+      expect(getAllByText("Api1").length).toEqual(2);
+      // Left pane active state
+      expect(
+        getByTestId("t--entity-item-Api1").classList.contains("active"),
+      ).toBe(true);
+      // Tabs active state
+      expect(getByTestId("t--ide-tab-api1").classList.contains("active")).toBe(
+        true,
+      );
+      // Check if the form is rendered
+      getByTestId("t--api-editor-form");
+      // Check if the params tabs is visible
+      getByRole("tab", { name: /params/i });
+      // Check if run button is visible
+      expect(getAllByRole("button", { name: /run/i })).toHaveLength(2);
+      // Check if the Add new button is shown
+      getByTestId("t--add-item");
+
+      // Check if the bottom view is rendered
+
+      getByRole("tab", { name: /response/i, selected: true });
+
+      expect(getAllByRole("tab", { name: /headers/i })).toHaveLength(2);
+      getByRole("tab", { name: /logs/i });
+      getByRole("tab", { name: /linter/i });
+    });
+
+    it("Renders Api routes in Split Screen", async () => {
+      const page = PageFactory.build();
+      const anApi = APIFactory.build({
+        id: "api_id2",
+        baseId: "api_base_id2",
+        pageId: page.pageId,
+      });
+      const state = getIDETestState({
+        actions: [anApi],
+        pages: [page],
+        tabs: {
+          [EditorEntityTab.QUERIES]: [anApi.baseId],
+          [EditorEntityTab.JS]: [],
+        },
+        ideView: EditorViewMode.SplitScreen,
+      });
+
+      const { getAllByRole, getAllByText, getByRole, getByTestId } = render(
+        <Route path={BUILDER_PATH}>
+          <IDE />
+        </Route>,
+        {
+          url: `/app/applicationSlug/pageSlug-${page.basePageId}/edit/api/${anApi.baseId}`,
+          initialState: state,
+        },
+      );
+
+      // Check if editor is in split screen
+      getByTestId("t--ide-maximize");
+      getByTestId("t--widgets-editor");
+
+      // Check if api is rendered in side by side
+      expect(getAllByText("Api2").length).toBe(1);
+      // Tabs active state
+      expect(getByTestId("t--ide-tab-api2").classList.contains("active")).toBe(
+        true,
+      );
+      // Check if the form is rendered
+      getByTestId("t--api-editor-form");
+      // Check if run button is visible
+      expect(getAllByRole("button", { name: /run/i }).length).toBe(2);
+      // Check if the Add new button is shown
+      getByTestId("t--ide-tabs-add-button");
+
+      // Check if the bottom view is rendered
+
+      getByRole("tab", {
+        name: /response/i,
+        selected: true,
+      });
+
+      expect(getAllByRole("tab", { name: /headers/i })).toHaveLength(2);
+    });
+
+    it("Renders Api add routes in Full Screen", () => {
+      const page = PageFactory.build();
+      const anApi = APIFactory.build({
+        id: "api_id",
+        baseId: "api_base_id",
+        pageId: page.pageId,
+      });
+      const state = getIDETestState({
+        actions: [anApi],
+        pages: [page],
+        tabs: {
+          [EditorEntityTab.QUERIES]: [anApi.baseId],
+          [EditorEntityTab.JS]: [],
+        },
+      });
+
+      const { getByTestId, getByText } = render(
+        <Route path={BUILDER_PATH}>
+          <IDE />
+        </Route>,
+        {
+          url: `/app/applicationSlug/pageSlug-${page.basePageId}/edit/api/${anApi.baseId}/add`,
+          initialState: state,
+        },
+      );
+
+      // Create options are rendered
+      getByText(createMessage(EDITOR_PANE_TEXTS.queries_create_from_existing));
+      getByText("New datasource");
+      getByText("REST API");
+      // Check new tab presence
+      const newTab = getByTestId("t--ide-tab-new_query");
+
+      expect(newTab).not.toBeNull();
+      // Close button is rendered
+      expect(
+        newTab.querySelector("[data-testid='t--tab-close-btn']"),
+      ).not.toBeNull();
+    });
+
+    it("Renders Api add routes in Split Screen", () => {
+      const page = PageFactory.build();
+      const anApi = APIFactory.build({
+        id: "api_id",
+        baseId: "api_base_id",
+        pageId: page.pageId,
+      });
+      const state = getIDETestState({
+        actions: [anApi],
+        pages: [page],
+        tabs: {
+          [EditorEntityTab.QUERIES]: [anApi.baseId],
+          [EditorEntityTab.JS]: [],
+        },
+        ideView: EditorViewMode.SplitScreen,
+      });
+
+      const { getAllByText, getByTestId, getByText, queryByTestId } = render(
+        <Route path={BUILDER_PATH}>
+          <IDE />
+        </Route>,
+        {
+          url: `/app/applicationSlug/pageSlug-${page.basePageId}/edit/api/${anApi.baseId}/add`,
+          initialState: state,
+        },
+      );
+
+      // There will be 1 Api4 text ( The tab )
+      expect(getAllByText("Api4").length).toEqual(1);
+      // Tabs active state
+      expect(getByTestId("t--ide-tab-api4").classList.contains("active")).toBe(
+        false,
+      );
+      // Add button should not present
+      expect(queryByTestId("t--ide-tabs-add-button")).toBeNull();
+
+      // Check if the form is not rendered
+      expect(queryByTestId("t--action-form-API")).toBeNull();
+      // Create options are rendered
+      getByText(createMessage(EDITOR_PANE_TEXTS.queries_create_from_existing));
+      getByText("New datasource");
+      getByText("REST API");
+      // Check new tab presence
+      const newTab = getByTestId("t--ide-tab-new_query");
+
+      expect(newTab).not.toBeNull();
+      // Close button is rendered
+      expect(
+        newTab.querySelector("[data-testid='t--tab-close-btn']"),
+      ).not.toBeNull();
+    });
+  });
+});
